+++ src/extension/notebook/httpNotebookKernel.ts
@@ -1,24 +1,21 @@
 import * as vscode from 'vscode';
 import type * as Httpyac from 'httpyac';
-import * as extensionApi from '../extensionApi';
-import * as httpOutput from '../httpOutputProvider';
-import { AppConfig, TestSlotOutput } from '../config';
 import { HttpNotebookViewType } from './notebookSelector';
+import { HttpNotebookOutputFactory } from './httpNotebookOutputFactory';
 
 
 export class HttpNotebookKernel implements vscode.NotebookCellStatusBarItemProvider {
 
   private executionOrder = 0;
 
-  readonly httpOutputProvider: Array<extensionApi.HttpOutputProvider>;
 
   private subscriptions: Array<vscode.Disposable>;
   onDidChangeCellStatusBarItems?: vscode.Event<void> | undefined;
   constructor(
+    private readonly httpNotebookOutputFactory: HttpNotebookOutputFactory,
     private readonly httpyac: typeof Httpyac,
     private readonly httpFileStore: Httpyac.HttpFileStore,
-    private readonly refreshCodeLens: vscode.EventEmitter<void>,
-    readonly config: AppConfig,
+    private readonly refreshCodeLens: vscode.EventEmitter<void>
   ) {
     this.onDidChangeCellStatusBarItems = refreshCodeLens.event;
     const controller = vscode.notebook.createNotebookController('httpbook-kernel', HttpNotebookViewType, 'HttpBook');
@@ -27,13 +24,6 @@ export class HttpNotebookKernel implements vscode.NotebookCellStatusBarItemProvi
     controller.description = 'a Notebook for sending REST, SOAP, and GraphQL requests';
     controller.executeHandler = this.send.bind(this);
 
-    this.httpOutputProvider = [
-      new httpOutput.TestResultsMimeOutpoutProvider(),
-      new httpOutput.MonacoEditorHttpOutputProvider(config, this.httpyac),
-      new httpOutput.Rfc7230HttpOutpoutProvider(config, this.httpyac),
-      new httpOutput.ImageHttpOutputProvider(),
-      new httpOutput.ContentTypeHttpOutputProvider(config),
-    ];
 
     this.subscriptions = [
       controller,
@@ -142,10 +132,8 @@ export class HttpNotebookKernel implements vscode.NotebookCellStatusBarItemProvi
       const outputs: Array<vscode.NotebookCellOutput> = [];
 
       for (const httpRegion of httpRegions) {
-        outputs.push(...this.createHttpRegionOutputs(httpRegion, {
-          cell,
+        outputs.push(...this.httpNotebookOutputFactory.createHttpRegionOutputs(httpRegion, {
           metaData: httpRegion.metaData,
-          httpRegion,
           mimeType: httpRegion.response?.contentType?.mimeType,
           httpFile
         }));
@@ -166,104 +154,5 @@ export class HttpNotebookKernel implements vscode.NotebookCellStatusBarItemProvi
       execution.end({ success: false, endTime: Date.now() });
       return false;
     }
-  }
-
-  private createHttpRegionOutputs(httpRegion: Httpyac.HttpRegion, httpOutputContext: extensionApi.HttpOutputContext): vscode.NotebookCellOutput[] {
-    const outputs: vscode.NotebookCellOutput[] = [];
-    if (httpRegion.testResults && this.canShowTestResults(httpRegion.testResults)) {
-      const testResults = httpRegion.testResults;
-      const outputItems = this.mapHttpOutputProvider(
-        obj => !!obj.getTestResultOutputResult && obj.getTestResultOutputResult(testResults, httpOutputContext)
-      );
-      if (outputItems.length > 0) {
-        outputs.push(this.createNotebookCellOutput(outputItems, httpRegion.response?.contentType?.mimeType));
-      }
-    }
-    if (httpRegion.response) {
-      const response = httpRegion.response;
-      const outputItems = this.mapHttpOutputProvider(
-        obj => !!obj.getResponseOutputResult && obj.getResponseOutputResult(response, httpOutputContext)
-      );
-      if (outputItems.length > 0) {
-        outputs.push(this.createNotebookCellOutput(outputItems, response.contentType?.mimeType));
-      }
-    }
-    return outputs;
-  }
-
-  private canShowTestResults(testResults: Array<Httpyac.TestResult> | undefined) {
-    if (testResults && this.config.outputTests !== TestSlotOutput.never) {
-      if (this.config.outputTests === TestSlotOutput.onlyFailed) {
-        return testResults.some(obj => !obj.result);
-      }
-      return true;
-    }
-    return false;
-  }
-
-  private mapHttpOutputProvider(mapFunc: (obj: extensionApi.HttpOutputProvider) => (extensionApi.HttpOutputResult | false)) {
-    const result: Array<extensionApi.HttpOutputResult> = [];
-
-    for (const httpOutputProvider of this.httpOutputProvider) {
-      try {
-        const obj = mapFunc(httpOutputProvider);
-        if (obj) {
-          result.push(obj);
-        }
-      } catch (err) {
-        this.httpyac.log.error(httpOutputProvider.id, err);
-      }
-    }
-    return result;
-  }
-
-  private createNotebookCellOutput(outputItems: Array<extensionApi.HttpOutputResult>, mimeType?: string) {
-    const preferredMime = this.getPreferredNotebookOutputRendererMime(mimeType);
-
-    const items = outputItems
-      .sort((obj1, obj2) => this.compareHttpOutputResults(obj1, obj2, preferredMime))
-      .reduce((prev: Array<vscode.NotebookCellOutputItem>, current) => {
-        if (Array.isArray(current.outputItems)) {
-          prev.push(...current.outputItems);
-        } else {
-          prev.push(current.outputItems);
-        }
-        return prev;
-      }, [])
-      .filter((obj, index, self) => self.indexOf(obj) === index);
-
-    return new vscode.NotebookCellOutput(items);
-  }
-
-  private getPreferredNotebookOutputRendererMime(mimeType?: string) {
-    if (mimeType && this.config.preferNotebookOutputRenderer) {
-      for (const [regex, mime] of Object.entries(this.config.preferNotebookOutputRenderer)) {
-        const regexp = new RegExp(regex, 'ui');
-        if (regexp.test(mimeType)) {
-          return mime;
-        }
-      }
-    }
-    return '';
-  }
-
-  private compareHttpOutputResults(obj1: extensionApi.HttpOutputResult, obj2: extensionApi.HttpOutputResult, mime?: string) {
-
-    if (mime) {
-      if (this.hasHttpOutputResultsMime(obj1, mime)) {
-        return -1;
-      }
-      if (this.hasHttpOutputResultsMime(obj2, mime)) {
-        return 1;
-      }
-    }
-    return obj2.priority - obj1.priority;
-  }
-
-  private hasHttpOutputResultsMime(obj: extensionApi.HttpOutputResult, mime?: string) : boolean {
-    if (Array.isArray(obj.outputItems)) {
-      return obj.outputItems.some(obj => obj.mime === mime);
-    }
-    return obj.outputItems.mime === mime;
   }
 }