+++ src/extension/notebook/index.ts
@@ -1,3 +1,4 @@
 export * from './environmentController';
 export * from './httpNotebookContentProvider';
 export * from './httpNotebookKernel';
+export * from './httpNotebookOutputFactory';